requirements.txt
customtkinter
pillow
pystray
pygame
requests
ddgs
